+++ configs/paths.php
@@ -26,8 +26,10 @@ $pathsConfig = [
     "tests" => __DIR__ . "/../tests/app",
     // The path to the temporary directory
     "tmp" => __DIR__ . "/../tmp",
-    // The framework's temporary directory
-    "tmp.framework" => __DIR__ . "/../tmp/framework",
+    // The framework's console temporary directory
+    "tmp.framework.console" => __DIR__ . "/../tmp/framework/console",
+    // The framework's HTTP temporary directory
+    "tmp.framework.http" => __DIR__ . "/../tmp/framework/http",
     // The path to the vendor (Composer) directory
     "vendor" => __DIR__ . "/../vendor",
     // The path to the view directory