+++ tests/app/project/http/ApplicationTestCase.php
@@ -3,9 +3,8 @@
  * Defines the HTTP application test case
  */
 namespace Project\HTTP;
+use Closure;
 use RDev\Applications\Application;
-use RDev\Applications\Bootstrappers\Dispatchers\IDispatcher;
-use RDev\Applications\Bootstrappers\IO\BootstrapperIO;
 use RDev\Framework\Tests\HTTP\ApplicationTestCase as BaseTestCase;
 
 class ApplicationTestCase extends BaseTestCase
@@ -24,17 +23,22 @@ class ApplicationTestCase extends BaseTestCase
     protected function setApplication()
     {
         /** @var Application $application */
-        /** @var BootstrapperIO $bootstrapperIO */
-        /** @var IDispatcher $bootstrapperDispatcher */
         require __DIR__ . "/../../../../bootstrap/start.php";
-        $httpBootstrapperClasses = require $application->getPaths()["configs"] . "/http/bootstrappers.php";
-        $bootstrapperIO->registerBootstrapperClasses($httpBootstrapperClasses);
-        $application->registerPreStartTask(function() use ($bootstrapperDispatcher, &$bootstrapperIO)
-        {
-            $bootstrapperDispatcher->dispatch(
-                $bootstrapperIO->read(BootstrapperIO::CACHED_HTTP_BOOTSTRAPPER_REGISTRY_FILE_NAME)
-            );
-        });
         $this->application = $application;
+
+        /**
+         * ----------------------------------------------------------
+         * Setup the bootstrappers
+         * ----------------------------------------------------------
+         *
+         * @var Closure $configureBootstrappers
+         */
+        $configureBootstrappers = require __DIR__ . "/../../../../bootstrap/configureBootstrappers.php";
+        $configureBootstrappers(
+            $this->application,
+            require $application->getPaths()["configs"] . "/http/bootstrappers.php",
+            false,
+            false
+        );
     }
 }