+++ CHANGELOG.md
@@ -4,13 +4,17 @@
 * Updated to RDev 0.5.0
 * Jumped to version 0.5.0 to match RDev version
 * Added this changelog
+* Added `bootstrap/configureBootstrappers.php`
 * Added changelog to .gitattributes `export-ignore`
+* Moved tmp/http/* to tmp/framework/http/*
+* Added tmp/framework/console
 
 <h3>Bootstrappers</h3>
-* Added code to register and dispatch bootstrappers in `bootstrap/console/start.php`, `bootstrap/http/start.php`, and `configs/application.php`
+* Added code to register and dispatch bootstrappers in `bootstrap/configureBootstrappers.php`
 
 <h3>Configs</h3>
-* Added `tmp.framework` to `configs/paths.php`
+* Added `tmp.framework.console` to `configs/paths.php`
+* Added `tmp.framework.http` to `configs/paths.php`
 * Renamed `compiledViews` path in `configs/paths.php` to `views.compiled`
 
 <h3>Console</h3>